Added: Help, completions and colors

+++ src/Controllers/Command.php
@@ -7,16 +7,19 @@ use App\Models\ContactManager;
 class Command
 {
     private ContactManager $contactManager;
+    private Help $helpController;
 
     public function __construct()
     {
         $this->contactManager = new ContactManager();
+        $this->helpController = new Help($this->contactManager);
+        $this->helpController->getCompletion();
     }
 
     public function loopAndDispatch(): void
     {
         while (true) {
-            $line = readline('> ');
+            $line = readline('Entrez votre commande (help, list, detail, create, delete, quit) : ');
 
             $line = trim($line);
 
@@ -33,7 +36,7 @@ class Command
         $commandArray = preg_split('/[\s,]+/', $command);
         switch ($commandArray[0]) {
             case 'help':
-                echo "Commande saisie : help" . PHP_EOL;
+                $this->helpController->help();
                 break;
             case 'list':
                 $this->list();
@@ -41,15 +44,19 @@ class Command
             case 'detail':
                 $id = $commandArray[1];
                 if (!ctype_digit($id)) {
-                    echo "$id n'est pas valide" . PHP_EOL;
+                    echo Help::color("$id n'est pas valide", 'red') . PHP_EOL . PHP_EOL;
                     break;
                 }
-                $this->detail($id);
+                if (array_key_exists($id, $this->contactManager->getContacts())) {
+                    $this->detail($id);
+                } else {
+                    echo Help::color("$id n'existe pas", 'red') . PHP_EOL . PHP_EOL;
+                }
                 break;
             case 'create':
 
                 if (count($commandArray) < 4) {
-                    echo "Il faut 3 arguments" . PHP_EOL;
+                    echo Help::color("Il faut 3 arguments", 'red') . PHP_EOL . PHP_EOL;
                     break;
                 }
 
@@ -60,19 +67,19 @@ class Command
 
                 $phoneNumber = array_pop($parameters);
                 if (!preg_match("/^[0-9]{1,20}$/", $phoneNumber)) {
-                    echo "$phoneNumber n'est pas un N° de téléphone valide" . PHP_EOL;
+                    echo Help::color("$phoneNumber n'est pas un N° de téléphone valide", 'red') . PHP_EOL . PHP_EOL;
                     break;
                 }
 
                 $email = array_pop($parameters);
                 if (!preg_match("/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/", $email)) {
-                    echo "$email n'est pas un eMail valide" . PHP_EOL;
+                    echo Help::color("$email n'est pas un eMail valide", 'red') . PHP_EOL . PHP_EOL;
                     break;
                 }
 
                 $name = implode(" ", $parameters);
                 if (!preg_match("/^[\p{L}][\p{L}\s'-]{1,149}$/u", $name)) {
-                    echo "$name n'est pas un nom valide" . PHP_EOL;
+                    echo Help::color("$name n'est pas un nom valide", 'red') . PHP_EOL . PHP_EOL;
                     break;
                 }
 
@@ -81,45 +88,59 @@ class Command
             case 'delete':
                 $id = $commandArray[1];
                 if (!ctype_digit($id)) {
-                    echo "$id n'est pas valide" . PHP_EOL;
+                    echo Help::color("$id n'est pas valide", 'red') . PHP_EOL . PHP_EOL;
                     break;
                 }
-                $this->delete($id);
+                if (array_key_exists($id, $this->contactManager->getContacts())) {
+                    $this->delete($id);
+                } else {
+                    echo Help::color("$id n'existe pas", 'red') . PHP_EOL . PHP_EOL;
+                }
                 break;
             case 'quit':
                 $this->quit();
                 break;
             default:
-                echo "Commande invalide : help" . PHP_EOL;
+                echo Help::color("Commande invalide : ", 'red') . $command . PHP_EOL . PHP_EOL;
                 break;
         }
+        readline_add_history($command);
     }
 
     public function list(): void
     {
-        echo "Commande saisie : list" . PHP_EOL;
+        //echo "Commande saisie : list" . PHP_EOL;
+        echo Help::color("Liste des contacts :", 'blue') . PHP_EOL . PHP_EOL;
+        echo Help::color("id, name, email, phone number", 'blue') . PHP_EOL . PHP_EOL;
         foreach ($this->contactManager->findAll() as $contact) {
-            echo $contact . PHP_EOL;
+            echo Help::color($contact, 'green') . PHP_EOL . PHP_EOL;
         }
     }
 
     public function detail(int $id): void
     {
-        echo "Commande saisie : detail $id" . PHP_EOL;
-        $contact = $this->contactManager->findById($id);
-        echo $contact . PHP_EOL;
+        //echo "Commande saisie : detail $id" . PHP_EOL;
+        //$contact = $this->contactManager->findById($id);
+        $contact = $this->contactManager->getContacts()[$id];
+        echo Help::color($contact, 'green') . PHP_EOL . PHP_EOL;
     }
 
     public function create(string $name, string $email, string $phoneNumber): void
     {
-        echo "Commande saisie : create $name $email $phoneNumber" . PHP_EOL;
-        $this->contactManager->insertContact($name, $email, $phoneNumber);
+        //echo "Commande saisie : create $name $email $phoneNumber" . PHP_EOL;
+        if ($this->contactManager->insertContact($name, $email, $phoneNumber)) {
+            echo Help::color("Contact $name $email $phoneNumber créé parfaitement !", 'green') . PHP_EOL . PHP_EOL;
+        };
+        $this->contactManager->findAll();
     }
 
     public function delete(int $id): void
     {
-        echo "Commande saisie : delete $id" . PHP_EOL;
-        $this->contactManager->deleteContact($id);
+        //echo "Commande saisie : delete $id" . PHP_EOL;
+        if ($this->contactManager->deleteContact($id)) {
+            echo Help::color("Contact supprimé avec succès !", 'green') . PHP_EOL . PHP_EOL;
+        };
+        $this->contactManager->findAll();
     }
 
     public function quit(): void
